Update Windows scripts with Enterprise entry point
The main class for Neo4j is specified in these two Powershell scripts,
yet it was missing for Enterprise. The branch for Enterprise is missing
from the public Neo4j repo as well, yet it is present in the Neo4j
Enterprise .zip file.

I suspect that these two lines are added by some build step in the
private Neo4j repository. A diff between the ONgDB packaging scripts and
the Neo4j Enterprise packaging scripts revealed that these two lines
were the only difference aside from license changes.

The entry point classname was first renamed in this commit:
https://github.com/neo4j/neo4j/commit/235af36e375bd3a9c7dbf1a5d816bfc9b2f0366a

It was then removed altogether in this commit:
https://github.com/neo4j/neo4j/commit/d6a0ba62e2e8a7123ad18aabe5029a27cc4803c5